requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
alembic=1.3.1
arrow=0.15.5
babel=2.8.0
blinker=1.4
bzip2=1.0.8
ca-certificates=2020.1.1
certifi=2019.11.28
cffi=1.14.0
chardet=3.0.4
click=7.0
cryptography=2.8
dominate=2.4.0
ffmpeg=4.2
ffmpeg-python=0.2.0
flask=1.1.1
flask-babel=1.0.0
flask-bootstrap=3.3.7.1
flask-login=0.4.1
flask-mail=0.9.1
flask-migrate=2.4.0
flask-moment=0.9.0
flask-sqlalchemy=2.4.1
flask-wtf=0.14.2
freetype=2.9.1
future=0.18.2
gmp=6.1.2
guess-language-spirit=0.5.3
idna=2.9
itsdangerous=1.1.0
jinja2=2.11.1
jpeg=9c
jwt=0.6.1
lame=3.100
libcxx=4.0.1
libcxxabi=4.0.1
libedit=3.1.20181209
libffi=3.2.1
libiconv=1.15
libopus=1.3
libpng=1.6.37
libtiff=4.1.0
libvpx=1.7.0
mako=1.1.1
markupsafe=1.1.1
ncurses=6.2
olefile=0.46
openh264=1.8.0
openssl=1.1.1
pillow=7.0.0
pip=20.0.2
psutil=5.7.0
pyautogui=0.9.48
pycparser=2.20
pymsgbox=1.0.7
pyobjc-core=6.1
pyobjc-framework-cocoa=6.1
pyobjc-framework-quartz=5.3
pyscreeze=0.1.26
python=3.8.1
python-dateutil=2.8.1
python-dotenv=0.10.5
python-editor=1.0.4
python_abi=3.8
pytweening=1.0.3
pytz=2019.3
readline=7.0
redis-py=3.4.1
requests=2.23.0
rq=1.2.2
rq-dashboard=0.6.1
setuptools=45.2.0
six=1.14.0
sqlalchemy=1.3.13
sqlite=3.31.1
tk=8.6.8
urllib3=1.25.8
visitor=0.1.3
werkzeug=0.16.1
wheel=0.34.2
wtforms=2.2.1
x264=1!152.20180806
xz=5.2.4
youtube-dl=2020.2.16
zlib=1.2.11
zstd=1.3.7
